MAJOR update to convert fully to pure C and .Call interface, plus numerous tidy-ups. Package now compiling with all checks (-Wall -Wextra -Werror -pedantic) enabled.

+++ src/Cache.h
@@ -1,9 +1,12 @@
 #ifndef CACHE_H
 #define CACHE_H
 
+#include <stdint.h>
+#include <stddef.h>
+
 extern uint32_t *hap_data;
 extern uint32_t **hap_locus;
-extern int32_t num_inds;
-extern int32_t hap_size;
+extern size_t num_inds;
+extern size_t hap_size;
 
 #endif
